Ongoing effort to reimplement core protocol of EFI into EFIKit.

+++ Private/efiSDK/gnuefi/reloc_x86_64.c
@@ -1,8 +1,8 @@
 /* reloc_x86_64.c - position independent x86_64 ELF shared object relocator
    Copyright (C) 1999 Hewlett-Packard Co.
-	Contributed by David Mosberger <davidm@hpl.hp.com>.
+    Contributed by David Mosberger <davidm@hpl.hp.com>.
    Copyright (C) 2005 Intel Co.
-	Contributed by Fenghua Yu <fenghua.yu@intel.com>.
+    Contributed by Fenghua Yu <fenghua.yu@intel.com>.
 
     All rights reserved.
 
@@ -40,59 +40,58 @@
 
 #include <elf.h>
 
-EFI_STATUS _relocate (long ldbase, Elf64_Dyn *dyn,
-		      EFI_HANDLE image EFI_UNUSED,
-		      EFI_SYSTEM_TABLE *systab EFI_UNUSED)
+EFI_STATUS _relocate(long ldbase, Elf64_Dyn *dyn, EFI_HANDLE image EFI_UNUSED, EFI_SYSTEM_TABLE *systab EFI_UNUSED)
 {
-	long relsz = 0, relent = 0;
-	Elf64_Rel *rel = 0;
-	unsigned long *addr;
-	int i;
+    long relsz = 0, relent = 0;
+    Elf64_Rel *rel = 0;
+    unsigned long *addr;
+    int i;
 
-	for (i = 0; dyn[i].d_tag != DT_NULL; ++i) {
-		switch (dyn[i].d_tag) {
-			case DT_RELA:
-				rel = (Elf64_Rel*)
-					((unsigned long)dyn[i].d_un.d_ptr
-					 + ldbase);
-				break;
+    for (i = 0; dyn[i].d_tag != DT_NULL; ++i)
+    {
+        switch (dyn[i].d_tag)
+        {
+        case DT_RELA:
+            rel = (Elf64_Rel *)((unsigned long long int)dyn[i].d_un.d_ptr + ldbase);
+            break;
 
-			case DT_RELASZ:
-				relsz = dyn[i].d_un.d_val;
-				break;
+        case DT_RELASZ:
+            relsz = dyn[i].d_un.d_val;
+            break;
 
-			case DT_RELAENT:
-				relent = dyn[i].d_un.d_val;
-				break;
+        case DT_RELAENT:
+            relent = dyn[i].d_un.d_val;
+            break;
 
-			default:
-				break;
-		}
-	}
+        default:
+            break;
+        }
+    }
 
-        if (!rel && relent == 0)
-                return EFI_SUCCESS;
+    if (!rel && relent == 0)
+        return EFI_SUCCESS;
 
-	if (!rel || relent == 0)
-		return EFI_LOAD_ERROR;
+    if (!rel || relent == 0)
+        return EFI_LOAD_ERROR;
 
-	while (relsz > 0) {
-		/* apply the relocs */
-		switch (ELF64_R_TYPE (rel->r_info)) {
-			case R_X86_64_NONE:
-				break;
+    while (relsz > 0)
+    {
+        /* apply the relocs */
+        switch (ELF64_R_TYPE(rel->r_info))
+        {
+        case R_X86_64_NONE:
+            break;
 
-			case R_X86_64_RELATIVE:
-				addr = (unsigned long *)
-					(ldbase + rel->r_offset);
-				*addr += ldbase;
-				break;
+        case R_X86_64_RELATIVE:
+            addr = (unsigned long *)(ldbase + rel->r_offset);
+            *addr += ldbase;
+            break;
 
-			default:
-				break;
-		}
-		rel = (Elf64_Rel*) ((char *) rel + relent);
-		relsz -= relent;
-	}
-	return EFI_SUCCESS;
+        default:
+            break;
+        }
+        rel = (Elf64_Rel *)((char *)rel + relent);
+        relsz -= relent;
+    }
+    return EFI_SUCCESS;
 }
